// frontend/src/api/generated/model/monitoringTargetViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AutomationCapability } from './automationCapability';
import type { Crop } from './crop';
import type { TargetKind } from './targetKind';

export interface MonitoringTargetViewModel {
  id: string;
  code: string;
  commonName: string;
  scientificName: string;
  kind: TargetKind;
  crop: Crop;
  automation: AutomationCapability;
}